requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
blinker==1.9.0
click==8.1.7
discord.py==2.4.0
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
pyserial==3.5
RPi.GPIO==0.7.1
Werkzeug==3.1.3
yarl==1.17.1
